# Conflicts:
#	Projects/VC2019/linkicc/linkicc.vcxproj
#	Projects/VC2019/tiffdiff/tiffdiff.vcxproj
#	Projects/VC2019/tifficc/tifficc.vcxproj
#	README.1ST
#	include/lcms2mt.h
#	include/lcms2mt_plugin.h
#	plugins/fast_float/src/fast_float_cmyk.c
#	plugins/fast_float/src/fast_float_internal.h
#	plugins/fast_float/src/fast_float_sup.c
#	plugins/fast_float/src/fast_float_tethra.c
#	plugins/fast_float/testbed/fast_float_testbed.c
#	src/cmscnvrt.c
#	src/cmsintrp.c
#	src/cmslut.c
#	src/cmsopt.c
#	src/cmsps2.c
#	src/cmssamp.c
#	src/lcms2_internal.h
#	testbed/testplugin.c
#	utils/jpgicc/jpgicc.c
#	utils/linkicc/linkicc.c
#	utils/psicc/psicc.c
#	utils/tificc/tifdiff.c
#	utils/tificc/tificc.c
#	utils/transicc/transicc.c

+++ src/cmsps2.c
@@ -551,7 +551,7 @@ void EmitNGamma(cmsContext ContextID, cmsIOHANDLER* m, cmsUInt32Number n, cmsTon
             _cmsIOPrintf(ContextID, m, "dup ");
         }
         else {
-            snprintf(buffer, sizeof(buffer), "%s%d", nameprefix, i);
+            snprintf(buffer, sizeof(buffer), "%s%d", nameprefix, (int) i);
 		    buffer[sizeof(buffer)-1] = '\0';
             Emit1Gamma(ContextID, m, g[i], buffer);
         }
